+++ src/constants.js
@@ -29,8 +29,6 @@ const CONSTANTS = {
     COLOR: "color",
     LUMINOSITY: "luminosity",
   },
-  // null scale results in auto scaling calculation - recommended
-  TOKEN_OFFSET: { position: { x: null, y: null }, scale: null },
   TO_RADIANS: Math.PI / 180,
   TRANSPARENCY_THRESHOLD: 254,
   MASK_DENSITY: 400,